+++ src/main/java/org/voyager/controller/ResourceController.java
@@ -44,12 +44,6 @@ class ResourceController {
     private SearchLocationService searchLocationService;
     @Autowired
     private LocationService locationService;
-    @Autowired
-    private RouteService routeService;
-    @Autowired
-    private AirportsService airportsService;
-    @Autowired
-    private DeltaService deltaService;
 
     private static final Logger LOGGER = LoggerFactory.getLogger(ResourceController.class);
 
@@ -76,74 +70,6 @@ class ResourceController {
         return searchLocationService.attribution();
     }
 
-    @GetMapping("/nearby-airports")
-    @Cacheable("nearbyAirportsCache")
-    public List<AirportDisplay> nearbyAirports(@RequestParam(LATITUDE_PARAM_NAME) Double latitude,
-                                               @RequestParam(LONGITUDE_PARAM_NAME) Double longitude,
-                                               @RequestParam(name = LIMIT_PARAM_NAME,defaultValue = "5") Integer limit,
-                                               @RequestParam(name = TYPE_PARAM_NAME, required = false) String typeString,
-                                               @RequestParam(name = AIRLINE_PARAM_NAME, required = false) String airlineString) {
-        Option<AirportType> airportType = ValidationUtils.resolveTypeString(typeString);
-        Option<Airline> airline = ValidationUtils.resolveAirlineString(airlineString);
-        return airportsService.getByDistance(latitude,longitude,limit,airportType,airline);
-    }
-
-    @GetMapping("/iata")
-    @Cacheable("iataCodesCache")
-    public List<String> getIataCodes(@RequestParam(name = TYPE_PARAM_NAME, required = false) String typeString) {
-        Option<AirportType> typeOptional = ValidationUtils.resolveTypeString(typeString);
-        if (typeOptional.isEmpty()) return airportsService.getIata();
-        return airportsService.getIataByType(typeOptional.get());
-    }
-
-    @GetMapping("/routes/{id}")
-    public RouteDisplay getRouteById(@PathVariable(name = "id") String idString) {
-        Integer id = ValidationUtils.validateAndGetInteger(ID_PATH_VAR_NAME,idString,false);
-        Option<RouteDisplay> routeDisplay = routeService.getRouteById(id);
-        if (routeDisplay.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND,
-                MessageConstants.buildResourceNotFoundForPathVariableMessage(ID_PATH_VAR_NAME,String.valueOf(id)));
-        return routeDisplay.get();
-    }
-
-    @PatchMapping("/routes/{id}")
-    public RouteDisplay patchRouteById(@RequestBody @Valid @NotNull RoutePatch routePatch, BindingResult bindingResult, @PathVariable(name = "id") String idString) {
-        ValidationUtils.validateRoutePatch(routePatch,bindingResult);
-        Integer id = ValidationUtils.validateAndGetInteger(ID_PATH_VAR_NAME,idString,false);
-        Option<RouteDisplay> routeDisplay = routeService.getRouteById(id);
-        if (routeDisplay.isEmpty()) throw new ResponseStatusException(HttpStatus.NOT_FOUND,
-                MessageConstants.buildResourceNotFoundForPathVariableMessage(ID_PATH_VAR_NAME,String.valueOf(id)));
-        return routeService.patch(routeDisplay.get(),routePatch);
-    }
-
-    @GetMapping("/routes")
-    public List<RouteDisplay> getRoutes(@RequestParam(name = AIRLINE_PARAM_NAME, required = false) String airlineString, @RequestParam(name = ORIGIN_PARAM_NAME, required = false) String origin, @RequestParam(name = DESTINATION_PARAM_NAME, required = false) String destination,@RequestParam(name = IS_ACTIVE_PARAM_NAME, required = false) Boolean isActive) {
-        Option<String> originOption = Option.none();
-        Option<String> destinationOption = Option.none();
-        if (StringUtils.isNotEmpty(origin)) originOption = Option.of(ValidationUtils.validateIataToUpperCase(origin,airportsService,ORIGIN_PARAM_NAME,true));
-        if (StringUtils.isNotEmpty(destination)) destinationOption = Option.of(ValidationUtils.validateIataToUpperCase(destination,airportsService,DESTINATION_PARAM_NAME,true));
-        Option<Airline> airlineOption = ValidationUtils.resolveAirlineString(airlineString);
-        if (isActive == null) return routeService.getRoutes(originOption,destinationOption,airlineOption);
-        return routeService.getActiveRoutes(originOption,destinationOption,airlineOption,isActive);
-    }
-
-    @GetMapping("/path/{origin}/to/{destination}")
-    public PathDisplay getRoutes(@PathVariable(name = ORIGIN_PARAM_NAME) String origin, @PathVariable(name = DESTINATION_PARAM_NAME) String destination,@RequestParam(name = EXCLUDE_PARAM_NAME, required = false) List<String> exclusionList) {
-        origin = ValidationUtils.validateIataToUpperCase(origin,airportsService,ORIGIN_PARAM_NAME,false);
-        destination = ValidationUtils.validateIataToUpperCase(destination,airportsService,DESTINATION_PARAM_NAME,false);
-        Set<String> exclusionSet = Set.of();
-        if (exclusionList != null) {
-            exclusionList.replaceAll(iata -> ValidationUtils.validateIataToUpperCase(iata, airportsService, EXCLUDE_PARAM_NAME, true));
-            exclusionSet = Set.copyOf(exclusionList);
-        }
-        return routeService.buildPathWithExclusions(origin,destination,exclusionSet);
-    }
-
-    @PostMapping("/routes")
-    public RouteDisplay addRoute(@RequestBody @Valid @NotNull RouteForm routeForm, BindingResult bindingResult) {
-        ValidationUtils.validateRouteForm(routeForm, bindingResult);
-        return routeService.save(routeForm);
-    }
-
     @GetMapping("/locations")
     public List<LocationDisplay> getLocations(@RequestParam(name = SOURCE_PROPERTY_NAME,required = false) String sourceString, @RequestParam(name = SOURCE_ID_PARAM_NAME, required = false) String sourceId) {
         if (StringUtils.isEmpty(sourceString) && StringUtils.isEmpty(sourceId)) return locationService.getLocations();
@@ -165,54 +91,5 @@ class ResourceController {
     public LocationDisplay addLocation(@RequestBody @Valid @NotNull LocationForm locationForm, BindingResult bindingResult) {
         ValidationUtils.validateLocationForm(locationForm, bindingResult);
         return locationService.save(locationForm);
-    }
-
-    @GetMapping("/airports/{iata}")
-    @Cacheable("iataCache")
-    public AirportDisplay getAirportByIata(@PathVariable(IATA_PARAM_NAME) String iata) {
-        LOGGER.debug(String.format("fetching uncached airport by iata code: %s",iata));
-        iata = ValidationUtils.validateIataToUpperCase(iata,airportsService,IATA_PARAM_NAME,false);
-        return airportsService.getByIata(iata);
-    }
-
-    @GetMapping("/airports")
-    @Cacheable("airportsCache")
-    public List<AirportDisplay> getAirports(@RequestParam(name = COUNTRY_CODE_PARAM_NAME, required = false) String countryCodeString,
-                                            @RequestParam(name = TYPE_PARAM_NAME, required = false) String typeString,
-                                            @RequestParam(name = AIRLINE_PARAM_NAME, required = false) String airlineString) {
-        if (countryCodeString != null) countryCodeString = ValidationUtils.validateAndGetCountryCode(countryCodeString);
-        Option<AirportType> airportType = ValidationUtils.resolveTypeString(typeString);
-        Option<Airline> airline = ValidationUtils.resolveAirlineString(airlineString);
-        return airportsService.getAll(Option.of(countryCodeString),airportType,airline);
-    }
-
-    @GetMapping("/delta")
-    public List<DeltaDisplay> getDeltas(@RequestParam(name = DELTA_STATUS_PARAM_NAME, required = false) List<String> statusStringList) {
-        if (statusStringList == null) return deltaService.getAll();
-        List<DeltaStatus> statusList = ValidationUtils.resolveDeltaStatusList(statusStringList);
-        return deltaService.getAllByStatusList(statusList);
-    }
-
-    @PostMapping("/delta")
-    public DeltaDisplay addDelta(@RequestBody @Valid @NotNull DeltaForm deltaForm, BindingResult bindingResult) {
-        ValidationUtils.validateDeltaForm(deltaForm, bindingResult);
-        return deltaService.save(deltaForm);
-    }
-
-    @GetMapping("/delta/{iata}")
-    public DeltaDisplay getDeltaByIata(@PathVariable(name = IATA_PARAM_NAME) String iata, @RequestParam(name = DELTA_STATUS_PARAM_NAME, required = false) List<String> statusStringList) {
-        iata = iata.toUpperCase();
-        if (deltaService.exists(iata)) return deltaService.getByIata(iata).get();
-        throw new ResponseStatusException(HttpStatus.NOT_FOUND,
-                MessageConstants.buildResourceNotFoundForPathVariableMessage(IATA_PARAM_NAME,iata));
-    }
-
-    @PatchMapping("/delta/{iata}")
-    public DeltaDisplay patchDeltaByIata(@RequestBody @Valid @NotNull DeltaPatch deltaPatch, BindingResult bindingResult, @PathVariable(name = "iata") String iata) {
-        ValidationUtils.validateDeltaPatch(deltaPatch,bindingResult);
-        iata = iata.toUpperCase();
-        if (deltaService.exists(iata)) return deltaService.patch(deltaService.getByIata(iata).get(),deltaPatch);
-        throw new ResponseStatusException(HttpStatus.NOT_FOUND,
-                MessageConstants.buildResourceNotFoundForPathVariableMessage(IATA_PARAM_NAME,iata));
     }
 }
