+++ src/main/java/org/voyager/model/entity/Country.java
@@ -1,0 +1,4 @@
+package org.voyager.model.entity;
+
+public class Country {
+}
